pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phylax"
version = "1.2.2"
description = "CI-native regression enforcement for LLM outputs."
readme = "README.md"

requires-python = ">=3.10"
authors = [
    {name = "Phylax Team"}
]
keywords = ["llm", "regression", "ci", "deterministic", "enforcement", "tracing"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
]
openai = [
    "openai>=1.0.0",
]
google = [
    "google-genai>=0.5.0",
]
groq = [
    "groq>=0.4.0",
]
mistral = [
    "mistralai>=0.1.0",
]
huggingface = [
    "huggingface_hub>=0.20.0",
]
ollama = [
    "ollama>=0.1.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.26.0",
    "requests>=2.31.0",
]
all = [
    "phylax[server,openai,google,groq,mistral,huggingface,ollama]",
]

[project.scripts]
phylax = "phylax.cli.main:main"

[project.urls]
Homepage = "https://github.com/xXMohitXx/Phylax"
Documentation = "https://github.com/xXMohitXx/Phylax#readme"
Repository = "https://github.com/xXMohitXx/Phylax"

[tool.setuptools.packages.find]
include = ["phylax*"]

[tool.setuptools.package-data]
phylax = [
    "ui/*.html",
    "ui/*.js",
    "ui/*.css",
    "assets/**/*",
]
